Featured Links and some UI updates/changes.
Added the ability to mark links as featured and also added another view
and menu item for Featured Links.  Made a bunch of changes to the UI in
terms of making sure that the Members Only menu button is active when
working on things in that area.  I also changed some yellow buttons to
blue for Edits.